+++ src/main/java/haru/command/AddEvent.java
@@ -1,3 +1,5 @@
+package haru.command;
+
 import haru.exception.HaruException;
 import haru.model.Event;
 import haru.model.TaskTime;
