+++ src/main/java/haru/command/CommandContext.java
@@ -1,3 +1,5 @@
+package haru.command;
+
 import haru.model.TaskList;
 import haru.ui.UI;
