+++ src/main/java/haru/command/MarkTask.java
@@ -1,3 +1,5 @@
+package haru.command;
+
 import haru.exception.HaruException;
 import haru.model.Task;
 import haru.model.TaskList;
