+++ src/main/java/haru/command/UnmarkTask.java
@@ -1,3 +1,5 @@
+package haru.command;
+
 import haru.exception.HaruException;
 import haru.model.Task;
 import haru.model.TaskList;
@@ -6,8 +8,8 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-public class DeleteTask extends Command {
-    public DeleteTask(CommandContext ctx) {
+public class UnmarkTask extends Command {
+    public UnmarkTask(CommandContext ctx) {
         super(new HashMap<>(Map.of("primary", "task number")), ctx);
     }
 
@@ -15,8 +17,8 @@ public class DeleteTask extends Command {
     public void execute() throws HaruException, IOException {
         TaskList taskList = this.getTaskList();
         int id = taskList.parseTaskId(this.getRequiredOption("primary"));
-        Task task = taskList.remove(id);
-        this.getUI().show("Okay~! I will delete this task:");
+        Task task = taskList.unmark(id);
+        this.getUI().show("Okay~! I will unmark this task as not done:");
         this.getUI().show(task.toString());
     }
 }
